requirements.txt
matplotlib==3.1.2
Flask_Cors==3.0.8
Flask==1.1.1
numpy==1.13.3
albumentations==0.5.2
Flask_SocketIO==4.3.1
tensorflow==2.0.0
pandas==0.25.3
seaborn==0.11.1
mlflow==1.14.1
opencv_python_headless==4.5.1.48
PyYAML==5.4.1
scikit_learn==0.24.1
